+++ src/main/java/com/stockmanagement/model/FactureVente.java
@@ -27,10 +27,10 @@ public class FactureVente {
     @Column(name = "Date")
     private Date date;
     @Column(name = "Total_HT")
-    private Float totalHT;
+    private double totalHT;
     @Column(name = "Total_TVA")
-    private Float totalTVA;
+    private double totalTVA;
     @Column(name = "Total_TTC")
-    private Float totalTTC;
+    private double totalTTC;
 
 }
